Agents/requirements.txt
pandas
numpy
scikit-learn
langchain
langgraph
openai
matplotlib
seaborn
fastapi
uvicorn
langchain-ollama
langgraph
python-dotenv